+++ sdk/nodejs/nrqlAlertCondition.ts
@@ -37,18 +37,18 @@ import * as utilities from "./utilities";
  * 
  * The `term` mapping supports the following arguments:
  * 
- *   * `duration` - (Required) In minutes, must be in the range of `1` to `120`, inclusive.
- *   * `operator` - (Optional) `above`, `below`, or `equal`.  Defaults to `equal`.
- *   * `priority` - (Optional) `critical` or `warning`.  Defaults to `critical`.
- *   * `threshold` - (Required) Must be 0 or greater.
- *   * `timeFunction` - (Required) `all` or `any`.
+ * - `duration` - (Required) In minutes, must be in the range of `1` to `120`, inclusive.
+ * - `operator` - (Optional) `above`, `below`, or `equal`. Defaults to `equal`.
+ * - `priority` - (Optional) `critical` or `warning`. Defaults to `critical`.
+ * - `threshold` - (Required) Must be 0 or greater.
+ * - `timeFunction` - (Required) `all` or `any`.
  * 
  * ## NRQL
  * 
  * The `nrql` attribute supports the following arguments:
  * 
- *   * `query` - (Required) The NRQL query to execute for the condition.
- *   * `sinceValue` - (Required) The value to be used in the `SINCE <X> MINUTES AGO` clause for the NRQL query. Must be between `1` and `20`.
+ * - `query` - (Required) The NRQL query to execute for the condition.
+ * - `sinceValue` - (Required) The value to be used in the `SINCE <X> MINUTES AGO` clause for the NRQL query. Must be between `1` and `20`.
  *
  * > This content is derived from https://github.com/terraform-providers/terraform-provider-newrelic/blob/master/website/docs/r/nrql_alert_condition.html.markdown.
  */
@@ -79,36 +79,15 @@ export class NrqlAlertCondition extends pulumi.CustomResource {
         return obj['__pulumiType'] === NrqlAlertCondition.__pulumiType;
     }
 
-    /**
-     * Set whether to enable the alert condition. Defaults to `true`.
-     */
     public readonly enabled!: pulumi.Output<boolean | undefined>;
     public readonly expectedGroups!: pulumi.Output<number | undefined>;
     public readonly ignoreOverlap!: pulumi.Output<boolean | undefined>;
-    /**
-     * The title of the condition
-     */
     public readonly name!: pulumi.Output<string>;
-    /**
-     * A NRQL query. See NRQL below for details.
-     */
     public readonly nrql!: pulumi.Output<outputs.NrqlAlertConditionNrql>;
-    /**
-     * The ID of the policy where this condition should be used.
-     */
     public readonly policyId!: pulumi.Output<number>;
-    /**
-     * Runbook URL to display in notifications.
-     */
     public readonly runbookUrl!: pulumi.Output<string | undefined>;
-    /**
-     * A list of terms for this condition. See Terms below for details.
-     */
     public readonly terms!: pulumi.Output<outputs.NrqlAlertConditionTerm[]>;
     public readonly type!: pulumi.Output<string | undefined>;
-    /**
-     * Possible values are `singleValue`, `sum`.
-     */
     public readonly valueFunction!: pulumi.Output<string | undefined>;
 
     /**
@@ -170,36 +149,15 @@ export class NrqlAlertCondition extends pulumi.CustomResource {
  * Input properties used for looking up and filtering NrqlAlertCondition resources.
  */
 export interface NrqlAlertConditionState {
-    /**
-     * Set whether to enable the alert condition. Defaults to `true`.
-     */
     readonly enabled?: pulumi.Input<boolean>;
     readonly expectedGroups?: pulumi.Input<number>;
     readonly ignoreOverlap?: pulumi.Input<boolean>;
-    /**
-     * The title of the condition
-     */
     readonly name?: pulumi.Input<string>;
-    /**
-     * A NRQL query. See NRQL below for details.
-     */
     readonly nrql?: pulumi.Input<inputs.NrqlAlertConditionNrql>;
-    /**
-     * The ID of the policy where this condition should be used.
-     */
     readonly policyId?: pulumi.Input<number>;
-    /**
-     * Runbook URL to display in notifications.
-     */
     readonly runbookUrl?: pulumi.Input<string>;
-    /**
-     * A list of terms for this condition. See Terms below for details.
-     */
     readonly terms?: pulumi.Input<pulumi.Input<inputs.NrqlAlertConditionTerm>[]>;
     readonly type?: pulumi.Input<string>;
-    /**
-     * Possible values are `singleValue`, `sum`.
-     */
     readonly valueFunction?: pulumi.Input<string>;
 }
 
@@ -207,35 +165,14 @@ export interface NrqlAlertConditionState {
  * The set of arguments for constructing a NrqlAlertCondition resource.
  */
 export interface NrqlAlertConditionArgs {
-    /**
-     * Set whether to enable the alert condition. Defaults to `true`.
-     */
     readonly enabled?: pulumi.Input<boolean>;
     readonly expectedGroups?: pulumi.Input<number>;
     readonly ignoreOverlap?: pulumi.Input<boolean>;
-    /**
-     * The title of the condition
-     */
     readonly name?: pulumi.Input<string>;
-    /**
-     * A NRQL query. See NRQL below for details.
-     */
     readonly nrql: pulumi.Input<inputs.NrqlAlertConditionNrql>;
-    /**
-     * The ID of the policy where this condition should be used.
-     */
     readonly policyId: pulumi.Input<number>;
-    /**
-     * Runbook URL to display in notifications.
-     */
     readonly runbookUrl?: pulumi.Input<string>;
-    /**
-     * A list of terms for this condition. See Terms below for details.
-     */
     readonly terms: pulumi.Input<pulumi.Input<inputs.NrqlAlertConditionTerm>[]>;
     readonly type?: pulumi.Input<string>;
-    /**
-     * Possible values are `singleValue`, `sum`.
-     */
     readonly valueFunction?: pulumi.Input<string>;
 }